Show points on a map using a hardcoded location
Now queries an area on the map, retrieves the points and displays on
the map. Not quite performant as it is drawing on top of the map,
but the alternative is to draw to the symbol layer.

TODO: Use a repository for location because the currennt implementation
is going to make me cry.